search_agent/requirements.txt
google-adk>=1.0.0
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.90.0
python-dotenv>=1.0.0
